requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gitdb==4.0.11
GitPython==3.1.41
pillow==10.4.0
PyJWT==2.9.0
setuptools==69.0.3
smmap==5.0.1
sqlparse==0.5.1
Unidecode==1.3.8